Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
PyJWT==2.8.0
Werkzeug==2.3.7
openpyxl==3.1.2
